Moved url into registry
