+++ apps/browser/src/state/registry.ts
@@ -1,7 +1,10 @@
-import { Store } from "redux";
+import { Store as ReduxStore } from "redux";
+import { fetchComments } from "@browser/actions/comment";
+import { dispatch } from "@browser/state/index";
+import filter from "@util/filter";
 
 export default class Registry {
-  static store: Store;
+  static store: ReduxStore;
 
   // eslint-ignore-next-line
   static dispatch(data: any) {
@@ -9,11 +12,25 @@ export default class Registry {
     Registry.store.dispatch(data);
   }
 
-  static getStore(): Store {
+  static getStore(): ReduxStore {
     return this.store;
   }
 
-  static set(store: Store) {
+  static set(store: ReduxStore) {
     this.store = store;
+  }
+}
+
+export class URL {
+  static filtered = "";
+
+  static update() {
+    const previous = this.filtered;
+
+    const filterResult = filter(window.location.href);
+    if (filterResult.error) return;
+    this.filtered = filterResult.url;
+
+    if (this.filtered !== previous) dispatch(fetchComments(0));
   }
 }
