+++ apps/browser/src/util/events.ts
@@ -3,11 +3,11 @@ import { checkLoggedIn } from "@browser/actions/user";
 import { Tab, Clipboard, Indicator } from "@browser/util/browser";
 import { Snackbar, Toast } from "@browser/util/dialog";
 import { dispatch } from "@browser/state/index";
+import { URL } from "@browser/state/registry";
 import Settings from "@browser/util/settings";
 import Domain from "@browser/util/domain";
 import Policy from "@browser/util/policy";
 import Render from "@browser/util/render";
-import URL from "@browser/util/url";
 import User from "@browser/util/user";
 import App from "@browser/util/app";
 type Data = { [key: string]: any };
